Final submission for SQL project
A program for a bookstore. The program allows the user to enter data about new books into the database, update book information, delete books from the database and search to find the availability of books in the database. This Level 2 Task 6 was submitted as part of the HyperionDev Data science bootcamp