Combining tables with SQL project
In this question 2 tables (collections of data) are given in the data.txt and keys.txt files. The goal of this question was to combine the tables based on the ages, sex and degrees of the individuals while also giving a count of the group and show the id of each person in that group. E.g (Agerang: 20, Count: 18, Id:2,11,36,42,44,45,51,52,57,65,67,68,69,79,86,88,90,91).